requirements.txt
# Core dependencies
pyautogen>=0.2.0  # AG2 framework
pydantic>=2.0.0
pyyaml>=6.0

# LLM providers
openai>=1.0.0
anthropic>=0.18.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Utilities
python-dotenv>=1.0.0
